start moving REPL widgets out of iframes (#2216)
